Prevent `"use cache"` timeout errors from being caught in userland code
When `dynamicIO` is enabled and a `"use cache"` function accesses
dynamic request APIs, we fail the prerendering with a timeout error
after 50 seconds.

This error could be swallowed in userland code however, when the cached
function is wrapped in a try/catch block. That's not the intended
behavior, so we now fail the prerendering (or dynamic validation in dev
mode) with the timeout error in this case as well, using the same
approach as in #77838.

This also works around a bug that led to the timeout errors not being
source-mapped correctly with Turbopack.

In a future PR, we will adapt the behaviour for prerendering of fallback
shells that are allowed to be empty, in which case the timeout must not
fail the build.